refactor: move ui out of client translation component
